requirements-docparse.txt
# Document parsing (PDF/tables/OCR)
pdfplumber>=0.11.0
pypdf>=4.0.0
pillow>=10.0.0
pytesseract>=0.3.10

